backend/requirements.txt
boto3==1.34.0
PyPDF2==3.0.1
fpdf2==2.7.6
fontTools==4.47.0
Pillow==10.1.0